# bcc-rosetta-server/src/server/openApi.json
{
  "openapi": "3.0.2",
  "info": {
    "version": "1.4.10",
    "title": "Rosetta",
    "description": "Build Once. Integrate Your Blockchain Everywhere.",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "paths": {
...
    "/block/transaction": {
      "post": {
        "summary": "Get a Block Transaction",
        "description": "Get a transaction in a block by its Transaction Identifier. This endpoint should only be used when querying a node for a block does not return all transactions contained within it. All transactions returned by this endpoint must be appended to any transactions returned by the /block method by consumers of this data. Fetching a transaction by hash is considered an Explorer Method (which is classified under the Future Work section). This method can be used to let consumers to paginate results when the  block trasactions count is too big to be returned in a single BlockResponse. Calling this endpoint requires reference to a BlockIdentifier because transaction parsing can change depending on which block contains the transaction. For example, in Bitcoin it is necessary to know which block contains a transaction to determine the destination of fee payments. Without specifying a block identifier, the node would have to infer which block to use (which could change during a re-org). Implementations that require fetching previous transactions to populate the response (ex: Previous UTXOs in Bitcoin) may find it useful to run a cache within the Rosetta server in the /data directory (on a path that does not conflict with the node).",
        "operationId": "blockTransaction",
        "tags": ["Block"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlockTransactionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Expected response to a valid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlockTransactionResponse"
                }
              }
            }
          },
          "500": {
            "description": "unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/account/balance": {
      "post": {
        "summary": "Get an Account's Balance",
        "description": "Get an array of all AccountBalances for an AccountIdentifier and the BlockIdentifier at which the balance lookup was performed. The BlockIdentifier must always be returned because some consumers of account balance data need to know specifically at which block the balance was calculated to compare balances they compute from operations with the balance returned by the node. It is important to note that making a balance request for an account without populating the SubAccountIdentifier should not result in the balance of all possible SubAccountIdentifiers being returned. Rather, it should result in the balance pertaining to no SubAccountIdentifiers being returned (sometimes called the liquid balance). To get all balances associated with an account, it may be necessary to perform multiple balance requests with unique AccountIdentifiers. It is also possible to perform a historical balance lookup (if the server supports it) by passing in an optional BlockIdentifier.",
        "operationId": "accountBalance",
        "tags": ["Account"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountBalanceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Expected response to a valid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountBalanceResponse"
                }
              }
            }
          },
          "500": {
            "description": "unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/account/coins": {
      "post": {
        "summary": "Get an Account's Unspent Coins",
        "description": "Get an array of all unspent coins for an AccountIdentifier and the BlockIdentifier at which the lookup was performed. If your implementation does not support coins (i.e. it is for an account-based blockchain), you do not need to implement this endpoint. If you implementation does support coins (i.e. it is fro a UTXO-based blockchain), you MUST also complete the `/account/balance` endpoint. It is important to note that making a coins request for an account without populating the SubAccountIdentifier should not result in the coins of all possible SubAccountIdentifiers being returned. Rather, it should result in the coins pertaining to no SubAccountIdentifiers being returned. To get all coins associated with an account, it may be necessary to perform multiple coin requests with unique AccountIdentifiers. Optionally, an implementation may choose to support updating an AccountIdentifier's unspent coins based on the contents of the mempool. Note, using this functionality breaks any guarantee of idempotency.",
        "operationId": "accountCoins",
        "tags": ["Account"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountCoinsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Expected response to a valid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountCoinsResponse"
                }
              }
            }
          },
          "500": {
            "description": "unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "NetworkIdentifier": {
        "description": "The network_identifier specifies which network a particular object is associated with.",
        "type": "object",
        "required": ["blockchain", "network"],
        "properties": {
          "blockchain": {
            "type": "string",
            "example": "bitcoin"
          },
          "network": {
            "description": "If a blockchain has a specific chain-id or network identifier, it should go in this field. It is up to the client to determine which network-specific identifier is mainnet or testnet.",
            "type": "string",
            "example": "mainnet"
          },
          "sub_network_identifier": {
            "$ref": "#/components/schemas/SubNetworkIdentifier"
          }
        }
      },
      "SubNetworkIdentifier": {
        "description": "In blockchains with sharded state, the SubNetworkIdentifier is required to query some object on a specific shard. This identifier is optional for all non-sharded blockchains.",
        "type": "object",
        "required": ["network"],
        "properties": {
          "network": {
            "type": "string",
            "example": "shard 1"
          },
          "metadata": {
            "type": "object",
            "example": {
              "producer": "0x52bc44d5378309ee2abf1539bf71de1b7d7be3b5"
            }
          }
        }
      },
      "BlockIdentifier": {
        "description": "The block_identifier uniquely identifies a block in a particular network.",
        "type": "object",
        "required": ["index", "hash"],
        "properties": {
          "index": {
            "description": "This is also known as the block height.",
            "type": "integer",
            "format": "int64",
            "example": 1123941
          },
          "hash": {
            "type": "string",
            "example": "0x1f2cc6c5027d2f201a5453ad1119574d2aed23a392654742ac3c78783c071f85"
          }
        }
      },
      "PartialBlockIdentifier": {
        "type": "object",
        "description": "When fetching data by BlockIdentifier, it may be possible to only specify the index or hash. If neither property is specified, it is assumed that the client is making a request at the current block.",
        "properties": {
          "index": {
            "type": "integer",
            "format": "int64",
            "example": 1123941
          },
          "hash": {
            "type": "string",
            "example": "0x1f2cc6c5027d2f201a5453ad1119574d2aed23a392654742ac3c78783c071f85"
          }
        }
      },
      "TransactionIdentifier": {
        "description": "The transaction_identifier uniquely identifies a transaction in a particular network and block or in the mempool.",
        "type": "object",
        "required": ["hash"],
        "properties": {
          "hash": {
            "description": "Any transactions that are attributable only to a block (ex: a block event) should use the hash of the block as the identifier.",
            "type": "string",
            "example": "0x2f23fd8cca835af21f3ac375bac601f97ead75f2e79143bdf71fe2c4be043e8f"
          }
        }
      },
...
      "AccountIdentifier": {
        "description": "The account_identifier uniquely identifies an account within a network. All fields in the account_identifier are utilized to determine this uniqueness (including the metadata field, if populated).",
        "type": "object",
        "required": ["address"],
        "properties": {
          "address": {
            "description": "The address may be a cryptographic public key (or some encoding of it) or a provided username.",
            "type": "string",
            "example": "0x3a065000ab4183c6bf581dc1e55a605455fc6d61"
          },
          "sub_account": {
            "$ref": "#/components/schemas/SubAccountIdentifier"
          },
          "metadata": {
            "description": "Blockchains that utilize a username model (where the address is not a derivative of a cryptographic public key) should specify the public key(s) owned by the address in metadata.",
            "type": "object"
          }
        }
      },
      "SubAccountIdentifier": {
        "description": "An account may have state specific to a contract address (ERC-20 token) and/or a stake (delegated balance). The sub_account_identifier should specify which state (if applicable) an account instantiation refers to.",
        "type": "object",
        "required": ["address"],
        "properties": {
          "address": {
            "description": "The SubAccount address may be a cryptographic value or some other identifier (ex: bonded) that uniquely specifies a SubAccount.",
            "type": "string",
            "example": "0x6b175474e89094c44da98b954eedeac495271d0f"
          },
          "metadata": {
            "description": "If the SubAccount address is not sufficient to uniquely specify a SubAccount, any other identifying information can be stored here. It is important to note that two SubAccounts with identical addresses but differing metadata will not be considered equal by clients.",
            "type": "object"
          }
        }
      },
...
      "Transaction": {
        "description": "Transactions contain an array of Operations that are attributable to the same TransactionIdentifier.",
        "type": "object",
        "required": ["transaction_identifier", "operations"],
        "properties": {
          "transaction_identifier": {
            "$ref": "#/components/schemas/TransactionIdentifier"
          },
          "operations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Operation"
            }
          },
          "related_transactions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RelatedTransaction"
            }
          },
          "metadata": {
            "description": "Transactions that are related to other transactions (like a cross-shard transaction) should include the tranaction_identifier of these transactions in the metadata.",
            "type": "object",
            "example": {
              "size": 12378,
              "lockTime": 1582272577
            }
          }
        }
      },
      "Operation": {
        "description": "Operations contain all balance-changing information within a transaction. They are always one-sided (only affect 1 AccountIdentifier) and can succeed or fail independently from a Transaction. Operations are used both to represent on-chain data (Data API) and to construct new transactions (Construction API), creating a standard interface for reading and writing to blockchains.",
        "type": "object",
        "required": ["operation_identifier", "type"],
        "properties": {
          "operation_identifier": {
            "$ref": "#/components/schemas/OperationIdentifier"
          },
          "related_operations": {
            "description": "Restrict referenced related_operations to identifier indices < the current operation_identifier.index. This ensures there exists a clear DAG-structure of relations. Since operations are one-sided, one could imagine relating operations in a single transfer or linking operations in a call tree.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OperationIdentifier"
            },
            "example": [
              {
                "index": 1
              },
              {
                "index": 2
              }
            ]
          },
          "type": {
            "description": "Type is the network-specific type of the operation. Ensure that any type that can be returned here is also specified in the NetworkOptionsResponse. This can be very useful to downstream consumers that parse all block data.",
            "type": "string",
            "example": "Transfer"
          },
          "status": {
            "description": "Status is the network-specific status of the operation. Status is not defined on the transaction object because blockchains with smart contracts may have transactions that partially apply (some operations are successful and some are not). Blockchains with atomic transactions (all operations succeed or all operations fail) will have the same status for each operation. On-chain operations (operations retrieved in the `/block` and `/block/transaction` endpoints) MUST have a populated status field (anything on-chain must have succeeded or failed). However, operations provided during transaction construction (often times called \"intent\" in the documentation) MUST NOT have a populated status field (operations yet to be included on-chain have not yet succeeded or failed).",
            "type": "string",
            "example": "Reverted"
          },
          "account": {
            "$ref": "#/components/schemas/AccountIdentifier"
          },
          "amount": {
            "$ref": "#/components/schemas/Amount"
          },
          "coin_change": {
            "$ref": "#/components/schemas/CoinChange"
          },
          "metadata": {
            "$ref": "#/components/schemas/OperationMetadata",
            "example": {
              "staking_credential": {
                "hex_bytes": "1B400D60AAF34EAF6DCBAB9BBA46001A23497886CF11066F7846933D30E5AD3F",
                "curve_type": "edwards25519"
              },
              "payment_key": {
                "hex_bytes": "1B400D60AAF34EAF6DCBAB9BBA46001A23497886CF11066F7846933D30E5AD3F",
                "curve_type": "edwards25519"
              },
              "asm": "304502201fd8abb11443f8b1b9a04e0495e0543d05611473a790c8939f089d073f90509a022100f4677825136605d732e2126d09a2d38c20c75946cd9fc239c0497e84c634e3dd01 03301a8259a12e35694cc22ebc45fee635f4993064190f6ce96e7fb19a03bb6be2",
              "hex": "48304502201fd8abb11443f8b1b9a04e0495e0543d05611473a790c8939f089d073f90509a022100f4677825136605d732e2126d09a2d38c20c75946cd9fc239c0497e84c634e3dd012103301a8259a12e35694cc22ebc45fee635f4993064190f6ce96e7fb19a03bb6be2"
            }
          }
        }
      },
...
      "Amount": {
        "description": "Amount is some Value of a Currency. It is considered invalid to specify a Value without a Currency.",
        "type": "object",
        "required": ["value", "currency"],
        "properties": {
          "value": {
            "description": "Value of the transaction in atomic units represented as an arbitrary-sized signed integer. For example, 1 BTC would be represented by a value of 100000000.",
            "type": "string",
            "example": "1238089899992"
          },
          "currency": {
            "$ref": "#/components/schemas/Currency"
          },
          "metadata": {
            "type": "object"
          }
        }
      },
...
      "Currency": {
        "description": "Currency is composed of a canonical Symbol and Decimals. This Decimals value is used to convert an Amount.Value from atomic units (Satoshis) to standard units (Bitcoins).",
        "type": "object",
        "required": ["symbol", "decimals"],
        "properties": {
          "symbol": {
            "description": "Canonical symbol associated with a currency.",
            "type": "string",
            "example": "BTC"
          },
          "decimals": {
            "description": "Number of decimal places in the standard unit representation of the amount. For example, BTC has 8 decimals. Note that it is not possible to represent the value of some currency in atomic units that is not base 10.",
            "type": "integer",
            "format": "int32",
            "minimum": 0,
            "example": 8
          },
          "metadata": {}
        }
      },
...
      "Coin": {
        "description": "Coin contains its unique identifier and the amount it represents.",
        "type": "object",
        "required": ["coin_identifier", "amount"],
        "properties": {
          "coin_identifier": {
            "$ref": "#/components/schemas/CoinIdentifier"
          },
          "amount": {
            "$ref": "#/components/schemas/Amount"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/TokenBundleItem"
              }
            }
          }
        }
      },
...
      "RelatedTransaction": {
        "description": "The related_transaction allows implementations to link together multiple transactions. An unpopulated network identifier indicates that the related transaction is on the same network.",
        "type": "object",
        "required": ["transaction_identifier", "direction"],
        "properties": {
          "network_identifier": {
            "$ref": "#/components/schemas/NetworkIdentifier"
          },
          "transaction_identifier": {
            "$ref": "#/components/schemas/TransactionIdentifier"
          },
          "direction": {
            "$ref": "#/components/schemas/Direction"
          }
        }
      },
...
      "AccountBalanceRequest": {
        "description": "An AccountBalanceRequest is utilized to make a balance request on the /account/balance endpoint. If the block_identifier is populated, a historical balance query should be performed.",
        "type": "object",
        "required": ["network_identifier", "account_identifier"],
        "properties": {
          "network_identifier": {
            "$ref": "#/components/schemas/NetworkIdentifier"
          },
          "account_identifier": {
            "$ref": "#/components/schemas/AccountIdentifier"
          },
          "block_identifier": {
            "$ref": "#/components/schemas/PartialBlockIdentifier"
          },
          "currencies": {
            "type": "array",
            "description": "In some cases, the caller may not want to retrieve all available balances for an AccountIdentifier. If the currencies field is populated, only balances for the specified currencies will be returned. If not populated, all available balances will be returned.",
            "items": {
              "$ref": "#/components/schemas/Currency"
            }
          }
        }
      },
      "AccountBalanceResponse": {
        "description": "An AccountBalanceResponse is returned on the /account/balance endpoint. If an account has a balance for each AccountIdentifier describing it (ex: an ERC-20 token balance on a few smart contracts), an account balance request must be made with each AccountIdentifier. The `coins` field was removed and replaced by by `/account/coins` in `v1.4.7`.",
        "type": "object",
        "required": ["block_identifier", "balances"],
        "properties": {
          "block_identifier": {
            "$ref": "#/components/schemas/BlockIdentifier"
          },
          "balances": {
            "type": "array",
            "description": "A single account may have a balance in multiple currencies.",
            "items": {
              "$ref": "#/components/schemas/Amount"
            }
          },
          "metadata": {
            "description": "Account-based blockchains that utilize a nonce or sequence number should include that number in the metadata. This number could be unique to the identifier or global across the account address.",
            "type": "object",
            "example": {
              "sequence_number": 23
            }
          }
        }
      },
      "AccountCoinsRequest": {
        "description": "AccountCoinsRequest is utilized to make a request on the /account/coins endpoint.",
        "type": "object",
        "required": ["network_identifier", "account_identifier"],
        "properties": {
          "network_identifier": {
            "$ref": "#/components/schemas/NetworkIdentifier"
          },
          "account_identifier": {
            "$ref": "#/components/schemas/AccountIdentifier"
          },
          "include_mempool": {
            "type": "boolean",
            "description": "Include state from the mempool when looking up an account's unspent coins. Note, using this functionality breaks any guarantee of idempotency."
          },
          "currencies": {
            "type": "array",
            "description": "In some cases, the caller may not want to retrieve coins for all currencies for an AccountIdentifier. If the currencies field is populated, only coins for the specified currencies will be returned. If not populated, all unspent coins will be returned.",
            "items": {
              "$ref": "#/components/schemas/Currency"
            }
          }
        }
      },
      "AccountCoinsResponse": {
        "description": "AccountCoinsResponse is returned on the /account/coins endpoint and includes all unspent Coins owned by an AccountIdentifier.",
        "type": "object",
        "required": ["block_identifier", "coins"],
        "properties": {
          "block_identifier": {
            "$ref": "#/components/schemas/BlockIdentifier"
          },
          "coins": {
            "type": "array",
            "description": "If a blockchain is UTXO-based, all unspent Coins owned by an account_identifier should be returned alongside the balance. It is highly recommended to populate this field so that users of the Rosetta API implementation don't need to maintain their own indexer to track their UTXOs.",
            "items": {
              "$ref": "#/components/schemas/Coin"
            }
          },
          "metadata": {
            "description": "Account-based blockchains that utilize a nonce or sequence number should include that number in the metadata. This number could be unique to the identifier or global across the account address.",
            "type": "object",
            "example": {
              "sequence_number": 23
            }
          }
        }
      },
...
      "BlockTransactionRequest": {
        "description": "A BlockTransactionRequest is used to fetch a Transaction included in a block that is not returned in a BlockResponse.",
        "type": "object",
        "required": ["network_identifier", "block_identifier", "transaction_identifier"],
        "properties": {
          "network_identifier": {
            "$ref": "#/components/schemas/NetworkIdentifier"
          },
          "block_identifier": {
            "$ref": "#/components/schemas/BlockIdentifier"
          },
          "transaction_identifier": {
            "$ref": "#/components/schemas/TransactionIdentifier"
          }
        }
      },
      "BlockTransactionResponse": {
        "description": "A BlockTransactionResponse contains information about a block transaction.",
        "type": "object",
        "required": ["transaction"],
        "properties": {
          "transaction": {
            "$ref": "#/components/schemas/Transaction"
          }
        }
      },
...
      "Error": {
        "description": "Instead of utilizing HTTP status codes to describe node errors (which often do not have a good analog), rich errors are returned using this object. Both the code and message fields can be individually used to correctly identify an error. Implementations MUST use unique values for both fields.",
        "type": "object",
        "required": ["code", "message", "retriable"],
        "properties": {
          "code": {
            "description": "Code is a network-specific error code. If desired, this code can be equivalent to an HTTP status code.",
            "type": "integer",
            "format": "int32",
            "minimum": 0,
            "example": 12
          },
          "message": {
            "description": "Message is a network-specific error message. The message MUST NOT change for a given code. In particular, this means that any contextual information should be included in the details field.",
            "type": "string",
            "example": "Invalid account format"
          },
          "description": {
            "description": "Description allows the implementer to optionally provide additional information about an error. In many cases, the content of this field will be a copy-and-paste from existing developer documentation. Description can ONLY be populated with generic information about a particular type of error. It MUST NOT be populated with information about a particular instantiation of an error (use `details` for this). Whereas the content of Error.Message should stay stable across releases, the content of Error.Description will likely change across releases (as implementers improve error documentation). For this reason, the content in this field is not part of any type assertion (unlike Error.Message).",
            "type": "string",
            "example": "This error is returned when the requested AccountIdentifier is improperly formatted."
          },
          "retriable": {
            "description": "An error is retriable if the same request may succeed if submitted again.",
            "type": "boolean"
          },
          "details": {
            "type": "object",
            "required": ["message"],
            "properties": {
              "message": { "type": "string" }
            },
            "description": "Often times it is useful to return context specific to the request that caused the error (i.e. a sample of the stack trace or impacted account) in addition to the standard error message.",
            "example": {
              "address": "0x1dcc4de8dec75d7aab85b567b6",
              "error": "not base64"
            }
          }
        }
      }
    }
  }
}